reworked the restart passenger stuff into a more generic refresh action
updated the script so it runs its commands using subprocess.

I updated the restart passenger action to check for a `config.ru`
file to know if it needs to run.  I also had it always make the
`tmp` dir so it would work even if that dir doesn't already exist.

I added a new action: `rebuild_jekyll_site` which will do just that
if your project folder has a `_config.yml` jekyll config file.  It
always runs with the `--no-auto` and `--no-server` flags so it will
only rebuild the site files in `_site`.